+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ This document describes changes between each past release.
 3.4.0 (unreleased)
 ==================
 
+**Breaking changes**
+
+- ``kinto --version`` was renamed ``kinto version``
+
 **New features**
 
 - Added a new built-in plugin ``kinto.plugins.history`` that keeps track of every action
@@ -16,6 +20,7 @@ This document describes changes between each past release.
 - Added ability to plug custom StatsD backend implementations via a new ``kinto.statsd_backend``
   setting. Useful for Datadog™ integration for example (fixes #626).
 - Added a ``delete-collection`` action to the ``kinto`` command. (#727)
+- Added verbosity options to the ``kinto`` command (#745)
 
 **Bug fixes**